Fix issue with duplicate resources(back to SageMaker in environment stack) - backwards compatibility
